fixes grid search (save/load models), new adj matrix algorithm

+++ code/luts.py
@@ -7,7 +7,7 @@ def lut_random(n_nodes, max_incoming_edges, p=0.5):
     # let idx represent the state of the reservoir: lut[0][idx] is the next state of node 0 with probability p
     assert 0 <= p <= 1
     lut = torch.rand((n_nodes, 2 ** max_incoming_edges)) < p
-    return lut
+    return lut.to(dtype=torch.uint8)
 
 if __name__ == '__main__':
     n = 100000
